move shuffle interfaces to org.apache.comet.shuffle package
